# frontend/openapi-recruitment.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conecta PRO - Módulo RECRUITMENT",
    "description": "API do módulo RECRUITMENT (Recrutamento e Seleção): vagas, candidatos, candidaturas, entrevistas",
    "version": "2.0.0"
  },
  "servers": [],
  "paths": {
    "/api/v1/recruitment/job-positions/": {
      "post": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Criar vaga",
        "description": "Cria uma nova vaga de emprego.",
        "operationId": "create_position_api_v1_recruitment_job_positions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPositionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Listar vagas",
        "description": "Lista vagas com filtros e paginação.",
        "operationId": "list_positions_api_v1_recruitment_job_positions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "position_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position Type"
            }
          },
          {
            "name": "position_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position Level"
            }
          },
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Department"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department"
            }
          },
          {
            "name": "work_model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkModel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Model"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "is_urgent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Urgent"
            }
          },
          {
            "name": "condominium_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Condominium Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Order By"
            }
          },
          {
            "name": "order_desc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Order Desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/open": {
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Listar vagas abertas",
        "description": "Lista vagas abertas para candidaturas.",
        "operationId": "list_open_positions_api_v1_recruitment_job_positions_open_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "condominium_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Condominium Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/expiring": {
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Vagas próximas da expiração",
        "description": "Lista vagas próximas da data limite.",
        "operationId": "list_expiring_positions_api_v1_recruitment_job_positions_expiring_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/stats": {
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Estatísticas de vagas",
        "description": "Retorna estatísticas das vagas.",
        "operationId": "get_position_stats_api_v1_recruitment_job_positions_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "condominium_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Condominium Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/{position_id}": {
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Buscar vaga",
        "description": "Busca vaga por ID.",
        "operationId": "get_position_api_v1_recruitment_job_positions__position_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Atualizar vaga",
        "description": "Atualiza uma vaga existente.",
        "operationId": "update_position_api_v1_recruitment_job_positions__position_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPositionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Remover vaga",
        "description": "Remove uma vaga (soft delete).",
        "operationId": "delete_position_api_v1_recruitment_job_positions__position_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/code/{code}": {
      "get": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Buscar vaga por código",
        "description": "Busca vaga por código.",
        "operationId": "get_position_by_code_api_v1_recruitment_job_positions_code__code__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/{position_id}/publish": {
      "post": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Publicar vaga",
        "description": "Publica uma vaga em rascunho.",
        "operationId": "publish_position_api_v1_recruitment_job_positions__position_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPositionPublish"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/job-positions/{position_id}/pause": {
      "post": {
        "tags": [
          "Recruitment - Recrutamento e Selecao",
          "Recruitment",
          "Recruitment - Vagas"
        ],
        "summary": "Pausar vaga",
        "description": "Pausa uma vaga aberta.",
        "operationId": "pause_position_api_v1_recruitment_job_positions__position_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PositionType": {
        "type": "string",
        "enum": [
          "clt",
          "pj",
          "temporario",
          "estagio",
          "trainee",
          "freelancer",
          "terceirizado"
        ],
        "title": "PositionType",
        "description": "Tipo de contratação."
      },
...
      "PositionLevel": {
        "type": "string",
        "enum": [
          "estagiario",
          "junior",
          "pleno",
          "senior",
          "especialista",
          "coordenador",
          "gerente",
          "diretor"
        ],
        "title": "PositionLevel",
        "description": "Nível da vaga."
      },
...
      "PositionStatus": {
        "type": "string",
        "enum": [
          "rascunho",
          "aberta",
          "pausada",
          "fechada",
          "cancelada",
          "preenchida"
        ],
        "title": "PositionStatus",
        "description": "Status da vaga."
      },
...
      "Department": {
        "type": "string",
        "enum": [
          "operacional",
          "administrativo",
          "financeiro",
          "rh",
          "comercial",
          "ti",
          "juridico",
          "marketing",
          "logistica",
          "seguranca",
          "facilities",
          "outro"
        ],
        "title": "Department",
        "description": "Departamento."
      },
...
      "WorkModel": {
        "type": "string",
        "enum": [
          "presencial",
          "remoto",
          "hibrido"
        ],
        "title": "WorkModel",
        "description": "Modelo de trabalho."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}